+++ Core/dataAccess/dataManager.py
@@ -27,4 +27,6 @@ def updateDataKey(dataLoc,keyToUpdate):
                     actionUnits[key] = None
 
     saveData(data,dataLoc)
+    
+